Remove `fields` helper from being tested.
Without this we get an error as it is tested like a normal helper.
This won't work as it needs a block. It should actually be tested
exactly like `fields_for` since it is a replacement for that helper.

I considered adding tests for this helper as well as the related
`form_with` replacement of `form_for`. Decided against this for
two reasons.

1. Although some of the `form_with`/`fields` helpers have been merged
   into Rails it is not done yet and subject to change.
   https://github.com/rails/rails/issues/25197
2. We are going to have to do more later anyway once this feature
   makes it's way to a release and it will get a bit messy as we'll
   have to skip tests using `form_with`/`fields` on older Rails.
   Rather than do a half-attempt now let's make a concentrated effort
   once the feature is finalized.