Remove deprecation warnings. Fix error due to mock not reflecting reality.
Checking to not raise a specific error has been deprecated in rspec. It is
suggested now to just have it test for no error as another type of error
will make it pass.

Turns out that deprecation suggestion is exactly on target as it was
hiding the fact that `permit` was being called on a plain hash instead
of a hash wrapped with the strong attributes support.

This fixes the error as well as removes that deprecation.